requirements-shared.txt
kombu
pydantic
pydantic-redis
pymongo
python-dotenv
redis
